rename bytes/file.py to bytes/local.py
