+++ dask/bytes/__init__.py
@@ -4,4 +4,4 @@ from .core import read_bytes
 with ignoring(ImportError):
     from . import s3
 with ignoring(ImportError):
-    from . import file
+    from . import local
